Incorporate fix43 into RoboCommand.cs
